rename: document -> docs로 워딩 수정
